Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main(){
-    int a, b, c, average;
-    scanf("%d", &a);
-    scanf("%d", &b);
-    scanf("%d", &c);
+    float a, b, c, average;
+    scanf("%f", &a);
+    scanf("%f", &b);
+    scanf("%f", &c);
     average=(a+b+c)/3;
     print("Average: %.2f", average);  
 }
